Split main function into a few smaller functions, move shared data to use header
